refactor #75 test accommodation
