+++ src/test/groovy/pl/sknikod/kodemy/infrastructure/rest/RoleServiceSpec.groovy
@@ -1,6 +1,7 @@
 package pl.sknikod.kodemy.infrastructure.rest
 
-import pl.sknikod.kodemy.infrastructure.model.entity.RoleName
+import pl.sknikod.kodemy.infrastructure.common.entity.RoleName
+import pl.sknikod.kodemy.infrastructure.user.RoleService
 import spock.lang.Specification
 
 class RoleServiceSpec extends Specification {
