+++ src/test/groovy/pl/sknikod/kodemy/infrastructure/rest/SectionServiceSpec.groovy
@@ -1,9 +1,10 @@
 package pl.sknikod.kodemy.infrastructure.rest
 
-import pl.sknikod.kodemy.infrastructure.model.entity.Section
-import pl.sknikod.kodemy.infrastructure.model.repository.SectionRepository
-import pl.sknikod.kodemy.infrastructure.rest.mapper.SectionMapper
-import pl.sknikod.kodemy.infrastructure.rest.model.SingleSectionResponse
+import pl.sknikod.kodemy.infrastructure.common.entity.Section
+import pl.sknikod.kodemy.infrastructure.common.repository.SectionRepository
+import pl.sknikod.kodemy.infrastructure.section.SectionService
+import pl.sknikod.kodemy.infrastructure.common.mapper.SectionMapper
+import pl.sknikod.kodemy.infrastructure.section.rest.SingleSectionResponse
 import spock.lang.Specification
 
 class SectionServiceSpec extends Specification {
