+++ src/test/groovy/pl/sknikod/kodemy/infrastructure/rest/UserControllerSpec.groovy
@@ -5,6 +5,8 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
 import pl.sknikod.kodemy.MvcIntegrationSpec
 import pl.sknikod.kodemy.WithKodemyMockUser
+import pl.sknikod.kodemy.infrastructure.user.UserService
+import pl.sknikod.kodemy.infrastructure.user.rest.UserController
 
 @WebMvcTest(UserController)
 class UserControllerSpec extends MvcIntegrationSpec {
